switching default spring package names to shorter version, removing excess imports
