+++ src/main/java/app/BuildingManagementApplication.java
@@ -1,0 +1,12 @@
+package app;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BuildingManagementApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(BuildingManagementApplication.class, args);
+	}
+}
