--- src/main/java/com/timsedam/buildingmanagement/BuildingManagementApplication.java
@@ -1,12 +1,0 @@
-package com.timsedam.buildingmanagement;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BuildingManagementApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(BuildingManagementApplication.class, args);
-	}
-}
