--- src/main/java/com/timsedam/buildingmanagement/model/Admin.java
@@ -1,10 +1,0 @@
-package com.timsedam.buildingmanagement.model;
-
-import javax.persistence.DiscriminatorValue;
-import javax.persistence.Entity;
-
-@Entity
-@DiscriminatorValue("admin")
-public class Admin extends User {
-	
-}
