--- src/main/java/com/timsedam/buildingmanagement/model/OpenEndedQuestion.java
@@ -1,11 +1,0 @@
-package com.timsedam.buildingmanagement.model;
-
-import java.util.List;
-
-import javax.persistence.Entity;
-import javax.persistence.OneToMany;
-
-@Entity
-public class OpenEndedQuestion extends Question {
-	
-}
