--- src/main/java/com/timsedam/buildingmanagement/model/Permission.java
@@ -1,46 +1,0 @@
-package com.timsedam.buildingmanagement.model;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@Entity
-public class Permission {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    public Long id;
-
-    @Column(unique = true)
-    public String name;
-
-    public Permission() {}
-
-    public Permission(String name) {
-        this.name = name;
-    }
-
-	public Long getId() {
-		return id;
-	}
-
-	public void setId(Long id) {
-		this.id = id;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
-
-	@Override
-	public String toString() {
-		return "Permission [id=" + id + ", name=" + name + "]";
-	}
-
-}
