--- src/main/java/com/timsedam/buildingmanagement/model/ProposalStatus.java
@@ -1,7 +1,0 @@
-package com.timsedam.buildingmanagement.model;
-
-public enum ProposalStatus {
-	OPEN,
-	ACCEPTED,
-	REJECTED
-}
