--- src/main/java/com/timsedam/buildingmanagement/model/ProposalVoteValue.java
@@ -1,6 +1,0 @@
-package com.timsedam.buildingmanagement.model;
-
-public enum ProposalVoteValue {
-	FOR,
-	AGAINST
-}
