--- src/main/java/com/timsedam/buildingmanagement/model/QuestionFormStatus.java
@@ -1,7 +1,0 @@
-package com.timsedam.buildingmanagement.model;
-
-public enum QuestionFormStatus {
-	OPEN,
-	STARTED,
-	CLOSED
-}
