--- src/main/java/com/timsedam/buildingmanagement/model/ReportStatus.java
@@ -1,7 +1,0 @@
-package com.timsedam.buildingmanagement.model;
-
-public enum ReportStatus {
-	OPEN,
-	PENDING,
-	CLOSED
-}
