--- src/main/java/com/timsedam/buildingmanagement/model/User.java
@@ -1,116 +1,0 @@
-package com.timsedam.buildingmanagement.model;
-
-import java.util.List;
-
-import javax.persistence.DiscriminatorColumn;
-import javax.persistence.DiscriminatorType;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Inheritance;
-import javax.persistence.InheritanceType;
-import javax.persistence.ManyToMany;
-import javax.persistence.OneToMany;
-
-@Entity
-@Inheritance(strategy = InheritanceType.SINGLE_TABLE)
-@DiscriminatorColumn(name = "user_type", discriminatorType = DiscriminatorType.STRING)
-public class User {
-
-	@Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-	private Long id;
-	private String username;
-	private String password;
-	private String email;
-	private String picture;   
-    @ManyToMany
-    private List<Role> roles;
-    @OneToMany(mappedBy = "reportCommented")
-    private List<Comment> comments;
-    
-	public User() {
-		super();
-	}
-
-	public User(String username, String password, String email, String picture, List<Role> roles,
-			List<Comment> comments) {
-		super();
-		this.username = username;
-		this.password = password;
-		this.email = email;
-		this.picture = picture;
-		this.roles = roles;
-		this.comments = comments;
-	}
-
-	public Long getId() {
-		return id;
-	}
-
-	public void setId(Long id) {
-		this.id = id;
-	}
-
-	public String getUsername() {
-		return username;
-	}
-
-	public void setUsername(String username) {
-		this.username = username;
-	}
-
-	public String getPassword() {
-		return password;
-	}
-
-	public void setPassword(String password) {
-		this.password = password;
-	}
-
-	public String getEmail() {
-		return email;
-	}
-
-	public void setEmail(String email) {
-		this.email = email;
-	}
-
-	public String getPicture() {
-		return picture;
-	}
-
-	public void setPicture(String picture) {
-		this.picture = picture;
-	}
-
-	public List<Role> getRoles() {
-		return roles;
-	}
-
-	public void setRoles(List<Role> roles) {
-		this.roles = roles;
-	}
-
-	public List<Comment> getComments() {
-		return comments;
-	}
-
-	public void setComments(List<Comment> comments) {
-		this.comments = comments;
-	}
-
-	public User(Long id, String username, String password, String email, String picture, List<Role> roles,
-			List<Comment> comments) {
-		super();
-		this.id = id;
-		this.username = username;
-		this.password = password;
-		this.email = email;
-		this.picture = picture;
-		this.roles = roles;
-		this.comments = comments;
-	}
-    
-}
