+++ src/main/java/model/Admin.java
@@ -1,0 +1,10 @@
+package model;
+
+import javax.persistence.DiscriminatorValue;
+import javax.persistence.Entity;
+
+@Entity
+@DiscriminatorValue("admin")
+public class Admin extends User {
+	
+}
