+++ src/main/java/model/OpenEndedQuestion.java
@@ -1,0 +1,8 @@
+package model;
+
+import javax.persistence.Entity;
+
+@Entity
+public class OpenEndedQuestion extends Question {
+	
+}
