+++ src/main/java/model/Permission.java
@@ -1,0 +1,47 @@
+package model;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+@Entity
+public class Permission {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    public Long id;
+    @Column(unique = true)
+    public String name;
+
+    public Permission() {
+		super();
+	}
+
+	public Permission(String name) {
+        this.name = name;
+    }
+
+	public Long getId() {
+		return id;
+	}
+
+	public void setId(Long id) {
+		this.id = id;
+	}
+
+	public String getName() {
+		return name;
+	}
+
+	public void setName(String name) {
+		this.name = name;
+	}
+
+	@Override
+	public String toString() {
+		return "Permission [id=" + id + ", name=" + name + "]";
+	}
+
+}
