+++ src/main/java/model/ProposalStatus.java
@@ -1,0 +1,7 @@
+package model;
+
+public enum ProposalStatus {
+	OPEN,
+	ACCEPTED,
+	REJECTED
+}
