+++ src/main/java/model/ProposalVoteValue.java
@@ -1,0 +1,6 @@
+package model;
+
+public enum ProposalVoteValue {
+	FOR,
+	AGAINST
+}
