+++ src/main/java/model/QuestionFormStatus.java
@@ -1,0 +1,7 @@
+package model;
+
+public enum QuestionFormStatus {
+	OPEN,
+	STARTED,
+	CLOSED
+}
