+++ src/main/java/model/ReportStatus.java
@@ -1,0 +1,7 @@
+package model;
+
+public enum ReportStatus {
+	OPEN,
+	PENDING,
+	CLOSED
+}
