+++ src/test/java/app/BuildingManagementApplicationTests.java
@@ -1,0 +1,16 @@
+package app;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class BuildingManagementApplicationTests {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
