--- src/test/java/com/timsedam/buildingmanagement/BuildingManagementApplicationTests.java
@@ -1,16 +1,0 @@
-package com.timsedam.buildingmanagement;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class BuildingManagementApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
